backend/requirements.txt
asgiref==3.8.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.3
django-allauth==64.2.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Markdown==3.10
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.8.0
python-dotenv==1.1.0
qrcode==8.2
requests==2.31.0
sqlparse==0.5.3
urllib3==2.6.3
